move som test utils functions into sui-core crate
