+++ crates/sui/tests/reconfiguration_tests.rs
@@ -118,7 +118,8 @@ use sui_types::messages_checkpoint::AuthenticatedCheckpoint;
 use sui_types::object::Object;
 use sui_types::{SUI_SYSTEM_STATE_OBJECT_ID, SUI_SYSTEM_STATE_OBJECT_SHARED_VERSION};
 use test_utils::authority::{get_object, start_node, test_authority_configs};
-use test_utils::messages::{make_transfer_sui_transaction, move_transaction};
+use test_utils::messages::move_transaction;
+use sui_core::test_utils::make_transfer_sui_transaction;
 use test_utils::objects::{generate_gas_object_with_balance, test_gas_objects};
 use test_utils::test_account_keys;
 use test_utils::transaction::{
